requirements.txt
asgiref==3.3.4
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-flex-fields==0.8.9
environ==1.0
factory-boy==3.2.0
Faker==8.1.2
Markdown==3.3.4
mysqlclient==2.0.3
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
